new: bold friends in tab list
Changes the formatting of friends in the tab menu to be bold.

The first time the player opens the tab menu, a thread will be started that sends a request to https://api.sk1er.club/friends/(playerUUID) to find a list of the friends. It will be stored in a List object.
From there, every time a player views a name in the tab menu, it will convert it to a username and check to see if it is in the friends list. If it is, the bold color code is added.