fix: bolding friend names not working in bedwars
Because of the way usernames are displayed in Bedwars, the option to
bold friend names in the tab list does not work. This problem is solved
by generalizing the way usernames are recognized and formatted.

In Bedwars, player names are formatted as §4§lR §r§4Steve, where R is
the team color and Steve is the username. Since the previous algorithm
didn’t work with this formatting, it is replaced with a regex string
that identifies if the first part of the name is a prefix.

Friend names are now formatted by adding §l after the last color code
of each word in the name.